delete useless file and move all to storage.go
