--- app/internal/storage/comments.go
@@ -1,31 +1,0 @@
-package storage
-
-import "CommentTree/internal/entities/comment"
-
-type db interface {
-	Create(c comment.Comment) (int64, error)
-	Comments(id int64) ([]comment.Comment, error)
-	Delete(id int64) error
-}
-
-type CommentStorage struct {
-	db
-}
-
-func NewComments(db db) *CommentStorage {
-	return &CommentStorage{
-		db: db,
-	}
-}
-
-func (cs *CommentStorage) Create(c comment.Comment) (int64, error) {
-	return 0, nil
-}
-
-func (cs *CommentStorage) Comments(id int64) ([]comment.Comment, error) {
-	return nil, nil
-}
-
-func (cs *CommentStorage) Delete(id int64) error {
-	return nil
-}
